backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.4",
    "django-cors-headers>=4.6.0",
    "django-extensions>=3.2.3",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
    "minio>=7.2.13",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
